Update platformio.ini
The used CPU has 4MB of flash memory. With the default settings only 1310720 bytes are used for the program
Current omote with default settings: Flash: [========= ]  85.1% (used 1115225 bytes from 1310720 bytes)

Using a different partition table will solve this problem. 

As the OTA feature is not needed this partition can be used for the application.
With this partition table 3145728 bytes are available for the application

Flash: [====      ]  35.5% (used 1115225 bytes from 3145728 bytes)