Update API class to better reuse parts of sending
Update Entity and light to reflect the API changes.

Add RapidJsonUtility for making it easier to debug and use for sending server updates

Curl does not seem to be working on hardware :(
So add http client (yet to be tested)

Still  in very new so may look to change API or base entity quite a bit.