update IHomeAssistApi in order to no longer do any parsing
Update downstream code to reflect new changes.

See HomeAssistApiParser to see an example of creating a RapidJSON SAX parser.
This version of the parser uses std::function to make it easily parse the pairs out. SAX is much more memory efficient since it only stores what is needed vs upfront parsing.

Ran into a crazy bug using default memory see SetBrightness() in the HomeAssist::Light class. RapidJSON default is pretty aggressive with memory allocation by default. This code keeps it all on the stack. Looking to make this better because this also has impact on which thread you can run this on since some threads don't have a lot of. Going to look at moving this on the heap for this reason. It would also be good to update the memPool to be a bit more transparent.  StackFrame size > task stack size = bad lol