Add a keypad gui to the simulator
This adds a new dependency on the SDL2 Image library.

It also adds a json parsing header libary to help load in the
information about the keys which we store in json.

The lvgl sdl driver does not filter SDL events by window ID so we have
to grab them with a filter and only pass the ones that aren't from our
windows.

I used this tool to draw the keypad collision polygons for future reference:
https://n-peugnet.github.io/image-map-creator/